pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.13.1",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "python-dotenv>=1.2.2",
    "smolagents[openai]==1.12.0",
]
